+++ src/main.cpp
@@ -31,7 +31,7 @@ SOFTWARE.
 #include <perf.hpp>
 #include <scale.hpp>
 #include <serialws.hpp>
-#include <temp.hpp>
+#include <temp_mgr.hpp>
 #include <utils.hpp>
 #include <wificonnection.hpp>
 #if CONFIG_IDF_TARGET_ESP32S2
@@ -48,7 +48,7 @@ KegPushHandler myPush(&myConfig);
 Display myDisplay;
 Scale myScale;
 LevelDetection myLevelDetection;
-TempHumidity myTemp;
+TempSensorManager myTemp;
 #if defined(USE_ASYNC_WEB)
 SerialWebSocket mySerialWebSocket;
 #endif
@@ -426,7 +426,7 @@ void logStartup() {
 #if defined(ESP8266)
     f.write(&s[0], strlen(&s[0]));
 #else
-    f.write((unsigned char*)&s[0], strlen(&s[0]));
+    f.write((unsigned char *)&s[0], strlen(&s[0]));
 #endif
     f.close();
   }